args to its own main
