+++ minori.py
@@ -1,25 +1,15 @@
 #!/usr/bin/env python3
 
-import sys
 import logging
 import shelve
-import click
 import feedparser
-
-logging.basicConfig(filename='minori.log',
-                    filemode='a',
-                    format='%(asctime)s [%(levelname)s] %(message)s',
-                    datefmt='%H:%M:%S',
-                    level=logging.DEBUG)
-logger = logging.getLogger('Minori')
-formatter = logging.Formatter('%(asctime)s [%(levelname)s] %(message)s')
-ch = logging.StreamHandler(sys.stdout)
-ch.setFormatter(formatter)
-logger.addHandler(ch)
 
 # TODO: move these later
 IN_PROGRESS = 0
 FINISHED = 1
+
+
+logger = logging.getLogger('Minori')
 
 
 class MinoriDatabase:
@@ -77,13 +67,26 @@ class MinoriDatabase:
                     logger.error("feed path did not lead to a valid dl string")
                     import pdb
                     pdb.set_trace()
+                    return
+            else:
+                # hope the rss feed follows the usual path of
+                # feed, entries, [x], link
+                dl_link = feed['entries'][0]['link']
+                if type(dl_link) != str:
+                    logger.error("no dl string found")
+                    import pdb
+                    pdb.set_trace()
+                    return
 
         except Exception as e:
             logger.error("Unable to parse feed, is it valid?")
+            import pdb
+            pdb.set_trace()
+            return
 
-        logger.info("Added feed!")
+        logger.info("Added feed {}".format(title))
 
-    def add_show(self, name, title_format, feed, max_eps=25, current_ep=0):
+    def add_show(self, name, title_format, feed, max_eps, current_ep):
         if feed not in self.db['feeds'].keys():
             logger.error("Feed {} not found, please add it first".format(feed))
             return
@@ -92,8 +95,8 @@ class MinoriDatabase:
                                   'feed': feed,
                                   'max_eps': max_eps,
                                   'current_ep': current_ep,
-                                  'status': IN_PROGRESS}
-        logger.info("Added show!")
+                                  'status': IN_PROGRESS if max_eps != current_ep else FINISHED}
+        logger.info("Added show {}".format(name))
 
     def check_for_shows():
         ''' 1. gather shows that are in progress
@@ -102,68 +105,3 @@ class MinoriDatabase:
             4. gather all links, and pass that over the downloader
         '''
         pass
-
-
-@click.group()
-@click.option('--db', default='db.shelve', help='name of db')
-@click.pass_context
-def cli(ctx, db):
-    if ctx.obj is None:
-        ctx.obj = {}
-    ctx.obj['db'] = db
-
-
-@cli.command()
-@click.pass_context
-def list_shows(ctx):
-    with MinoriDatabase(ctx.obj['db']) as m:
-        logger.info("Shows:")
-        logger.info(m.get_shows())
-
-
-@cli.command()
-@click.argument('name')
-@click.argument('title_format')
-@click.argument('feed')
-@click.option('--max-eps',  default=-1, help='max # of eps if known, default is -1')
-@click.option('--current-ep', default=0, help='episode currently on, e.g default 0 (not started)')
-@click.pass_context
-def add_show(ctx, name, title_format, feed, max_eps, current_ep):
-    with MinoriDatabase(ctx.obj['db']) as m:
-        m.add_show(name, title_format, feed, max_eps=max_eps, current_ep=current_ep)
-
-
-@cli.command()
-@click.pass_context
-def list_feeds(ctx):
-    with MinoriDatabase(ctx.obj['db']) as m:
-        logger.info("Feeds:")
-        logger.info(m.get_feeds())
-
-
-@cli.command()
-@click.argument('title')
-@click.argument('url')
-@click.argument('dl_command')
-@click.option('--feed-path', default='', help='advanced usage: feedparser path to link')
-@click.pass_context
-def add_feed(ctx, title, url, dl_command, feed_path):
-    with MinoriDatabase(ctx.obj['db']) as m:
-        m.add_feed(title, url, dl_command, feed_path)
-
-
-@cli.command()
-@click.pass_context
-def list_dl_commands(ctx):
-    with MinoriDatabase(ctx.obj['db']) as m:
-        logger.info("Commands:")
-        logger.info(m.get_dl_commands())
-
-
-@cli.command()
-@click.pass_context
-@click.argument('name')
-@click.argument('dl_command')
-def add_dl_command(ctx, name, dl_command):
-    with MinoriDatabase(ctx.obj['db']) as m:
-        m.add_dl_command(name, dl_command)
